@@ -17,6 +17,7 @@ from collections import defaultdict
 from contextlib import contextmanager
 import operator as op
 import threading
+from typing import Any, Callable, Dict, List, Sequence, Set, Type
 
 from absl import logging
 import numpy as onp
@@ -45,7 +46,7 @@ _map = safe_map
 
 def identity(x): return x
 
-def shard_args(backend, devices, assignments, axis_size, tuple_args, args):
+def shard_args(backend, devices, assignments, axis_size, args):
   """Shard each argument data array along its leading axis.
 
   Args:
@@ -96,14 +97,10 @@ def shard_args(backend, devices, assignments, axis_size, tuple_args, args):
       for r, buf in enumerate(bufs):
         buffers[r][a] = buf
 
-  if tuple_args:
-    buffers = [[xla.make_tuple(bufs, devices[r], backend)]
-               for r, bufs in enumerate(buffers)]
-
   return buffers
 
 
-shard_arg_handlers = {}
+shard_arg_handlers: Dict[Any, Callable[[Any, Any, Any], Sequence[Any]]] = {}
 shard_arg_handlers[core.Unit] = \
     lambda x, devices, _: [xla.device_put(core.unit, d) for d in devices]
 def _shard_array(x, devices, assignments):
@@ -125,7 +122,7 @@ def shard_aval(size, aval):
   except KeyError as err:
     raise TypeError("No shard_aval handler for type: {}".format(type(aval))
                     ) from err
-shard_aval_handlers = {}
+shard_aval_handlers: Dict[Type[core.AbstractValue], Callable[[int, Any], Any]] = {}
 shard_aval_handlers[core.AbstractUnit] = lambda size, x: x
 def _shard_abstract_array(size, x):
   if x.shape[0] != size:
@@ -140,7 +137,8 @@ def aval_to_result_handler(size, nrep, aval):
   except KeyError as err:
     raise TypeError("No pxla_result_handler for type: {}".format(type(aval))
                     ) from err
-pxla_result_handlers = {}
+PxlaResultHandler = Callable[..., Callable[[Any], Any]]
+pxla_result_handlers: Dict[Type[core.AbstractValue], PxlaResultHandler] = {}
 pxla_result_handlers[core.AbstractUnit] = lambda *_: lambda _: core.unit
 def array_result_handler(size, nrep, aval):
   full_aval = ShapedArray((size,) + aval.shape, aval.dtype)
@@ -246,7 +244,7 @@ def apply_parallel_primitive(prim, *args, **params):
     shape = (logical_size(dynamic_axis_env[axis_name]),)
   return parallel_pure_rules[prim](*args, shape=shape, **params)
 
-parallel_pure_rules = {}
+parallel_pure_rules: Dict[core.Primitive, Callable] = {}
 
 
 def axis_index(axis_name):
@@ -381,7 +379,7 @@ xla.canonicalize_dtype_handlers[ShardedDeviceArray] = identity
 
 
 class ChunkedDeviceArray(ShardedDeviceArray):
-  __slots__ = []
+  __slots__: List[str] = []
   _collect = staticmethod(onp.concatenate)
 
   def __init__(self, axis_size, aval, device_buffers):
@@ -546,13 +544,14 @@ def parallel_callable(fun, backend, axis_name, axis_size, global_axis_size,
 
   handle_args = partial(shard_args, backend, compiled.local_devices(),
                         assign_shards_to_replicas(num_local_replicas, axis_size),
-                        axis_size, tuple_args)
+                        axis_size)
   handle_outs = _pvals_to_results_handler(axis_size, num_local_replicas,
                                           out_pvals, compiled.local_devices(),
                                           backend)
-  return partial(execute_replicated, compiled, backend, handle_args, handle_outs)
+  return partial(execute_replicated, compiled, backend, handle_args, handle_outs,
+                 tuple_args)
 
-multi_host_supported_collectives = set()
+multi_host_supported_collectives: Set[core.Primitive] = set()
 
 class ResultToPopulate(object): pass
 result_to_populate = ResultToPopulate()
@@ -564,7 +563,7 @@ def _pvals_to_results_handler(size, nrep, out_pvals, devices, backend):
   def handler(out_bufs):
     buffers = [[result_to_populate] * nrep for _ in range(nouts)]
     for r, tuple_buf in enumerate(out_bufs):
-      for i, buf in enumerate(tuple_buf.destructure()):
+      for i, buf in enumerate(tuple_buf):
         buffers[i][r] = buf
     assert not any(buf is result_to_populate for bufs in buffers
                    for buf in bufs)
@@ -631,9 +630,11 @@ def _pval_to_result_handler(axis_size, nrep, pval, devices, backend):
   else:
     return aval_to_result_handler(axis_size, nrep, pv)
 
-def execute_replicated(compiled, backend, in_handler, out_handler, *args):
+def execute_replicated(compiled, backend, in_handler, out_handler, tuple_args,
+                       *args):
   input_bufs = in_handler(args)
-  out_bufs = compiled.ExecuteOnLocalDevices(list(input_bufs))
+  out_bufs = compiled.ExecuteOnLocalDevices(
+      list(input_bufs), tuple_arguments=tuple_args)
   return out_handler(out_bufs)
 
 
@@ -647,8 +648,7 @@ xla_pmap_p.def_impl(xla_pmap_impl)
 def _pmap_translation_rule(c, axis_env,
                            in_nodes, name_stack, axis_name, axis_size,
                            global_axis_size, devices, name,
-                           call_jaxpr, backend=None,
-                           mapped_invars=None):
+                           call_jaxpr, *, backend=None, mapped_invars):
   # We in-line here rather than generating a Call HLO as in the xla_call
   # translation rule just because the extra tuple stuff is a pain.
   if axis_env.devices is not None or (axis_env.names and devices is not None):
@@ -657,45 +657,48 @@ def _pmap_translation_rule(c, axis_env,
     global_axis_size = axis_size
   new_env = xla.extend_axis_env(axis_env, axis_name, global_axis_size)
   # Shard the in_nodes that are mapped
+  in_avals = [v.aval for v in call_jaxpr.invars]
   in_nodes_sharded = (
-    _xla_shard(c, new_env, in_node) if in_node_mapped else in_node
-    for in_node, in_node_mapped in zip(in_nodes, mapped_invars))
+    _xla_shard(c, aval, new_env, in_node) if in_node_mapped else in_node
+    for aval, in_node, in_node_mapped in zip(in_avals, in_nodes, mapped_invars))
 
   sharded_outs = xla.jaxpr_subcomp(
       c, call_jaxpr, backend, new_env, (),
       extend_name_stack(name_stack, wrap_name(name, 'pmap')), *in_nodes_sharded)
-  outs = [_xla_unshard(c, new_env, shard) for shard in sharded_outs]
+  out_avals = [v.aval for v in call_jaxpr.outvars]
+  outs = [_xla_unshard(c, aval, new_env, shard)
+          for aval, shard in zip(out_avals, sharded_outs)]
   return c.Tuple(*outs)
 
 xla.call_translations[xla_pmap_p] = _pmap_translation_rule
 ad.primitive_transposes[xla_pmap_p] = partial(ad.map_transpose, xla_pmap_p)
 pe.map_primitives.add(xla_pmap_p)
 
-def _xla_shard(c, axis_env, x):
-  xla_shape = c.GetShape(x)
-  if xla_shape.is_tuple():
-    assert not xla_shape.tuple_shapes()
+def _xla_shard(c, aval, axis_env, x):
+  if aval is core.abstract_unit:
     return x
-  else:
-    dims = list(xla_shape.dimensions())
+  elif isinstance(aval, ShapedArray):
+    dims = list(c.GetShape(x).dimensions())
     zero = c.Constant(onp.zeros((), dtype=onp.uint32))
     idxs = [_unravel_index(c, axis_env)] + [zero] * (len(dims) - 1)
     return c.Reshape(c.DynamicSlice(x, idxs, [1] + dims[1:]), None, dims[1:])
+  else:
+    raise TypeError((aval, c.GetShape(x)))
 
 # TODO(b/110096942): more efficient gather
-def _xla_unshard(c, axis_env, x):
-  xla_shape = c.GetShape(x)
-  if xla_shape.is_tuple():
-    assert not xla_shape.tuple_shapes()
+def _xla_unshard(c, aval, axis_env, x):
+  if aval is core.abstract_unit:
     return x
-  else:
-    dims = list(xla_shape.dimensions())
-    padded = c.Broadcast(c.Constant(onp.array(0, xla_shape.numpy_dtype())),
-                         [axis_env.sizes[-1]] + dims)
+  elif isinstance(aval, ShapedArray):
+    dims = list(c.GetShape(x).dimensions())
+    padded = c.Broadcast(c.Constant(onp.array(0, aval.dtype)),
+                          [axis_env.sizes[-1]] + dims)
     zero = c.Constant(onp.zeros((), dtype=onp.uint32))
     idxs = [_unravel_index(c, axis_env)] + [zero] * len(dims)
     padded = c.DynamicUpdateSlice(padded, c.Reshape(x, None, [1] + dims), idxs)
     return c.CrossReplicaSum(padded, xla.axis_groups(axis_env, axis_env.names[-1]))
+  else:
+    raise TypeError((aval, c.GetShape(x)))
 
 def _unravel_index(c, axis_env):
   div = c.Constant(onp.array(axis_env.nreps // prod(axis_env.sizes), onp.uint32))
@@ -864,4 +867,4 @@ class SplitAxisTrace(core.Trace):
     return  val, todo
 
 
-split_axis_rules = {}
+split_axis_rules: Dict[core.Primitive, Callable] = {}
