@@ -13,6 +13,7 @@
 # limitations under the License.
 
 from functools import partial
+from typing import Callable, Dict, Type
 
 from absl import logging
 import numpy as onp
@@ -86,7 +87,7 @@ def _pvals_to_results_handler(nrep, npar, partitions, out_pvals):
     buffers = [[[None] * npar for _ in range(nrep)] for _ in range(nouts)]
     for raw_idx, tuple_buf in enumerate(out_bufs):
       r, p = onp.unravel_index(raw_idx, (nrep, npar))
-      for i, buf in enumerate(tuple_buf.destructure()):
+      for i, buf in enumerate(tuple_buf):
         buffers[i][r][p] = buf
     return [h(bufs) for h, bufs in zip(handlers, buffers)]
 
@@ -105,7 +106,7 @@ def _aval_to_result_handler(partition, aval):
   return result_handlers[type(aval)](partition, aval)
 
 
-result_handlers = {}
+result_handlers: Dict[Type[core.AbstractValue], Callable] = {}
 
 
 def _array_result_handler(partition, aval):
@@ -202,7 +203,8 @@ def _sharded_jit_translation_rule(c, axis_env, freevar_nodes,
 
 def _execute_spatially_partitioned(compiled, in_handler, out_handler, *args):
   input_bufs = in_handler(args)
-  out_bufs = compiled.ExecuteOnLocalDevices(list(input_bufs))
+  out_bufs = compiled.ExecuteOnLocalDevices(
+      list(input_bufs), tuple_arguments=False)
   return out_handler(out_bufs)
 
 
