@@ -16,11 +16,11 @@ from .lax import *
 from .lax import (_reduce_sum, _reduce_max, _reduce_min, _reduce_or,
                   _reduce_and, _reduce_window_sum, _reduce_window_max,
                   _reduce_window_min, _reduce_window_prod,
-                  _select_and_gather_add, _float, _complex,
-                  _input_dtype, _const, _eq_meet, _safe_mul,
-                  _broadcasting_select, _check_user_dtype_supported,
-                  _one, _const, _upcast_fp16_for_computation,
-                  _broadcasting_shape_rule, _eye, _tri, _delta)
+                  _select_and_gather_add, _float, _complex, _input_dtype,
+                  _const, _eq_meet, _broadcasting_select,
+                  _check_user_dtype_supported, _one, _const,
+                  _upcast_fp16_for_computation, _broadcasting_shape_rule,
+                  _eye, _tri, _delta)
 from .lax_control_flow import *
 from .lax_fft import *
 from .lax_parallel import *
