@@ -113,13 +113,19 @@ def logsumexp(a, axis=None, b=None, keepdims=False, return_sign=False):
 @_wraps(osp_special.xlogy)
 def xlogy(x, y):
   x, y = _promote_args_inexact("xlogy", x, y)
-  return lax._safe_mul(x, lax.log(y))
+  x_ok = x != 0.
+  safe_x = jnp.where(x_ok, x, 1.)
+  safe_y = jnp.where(x_ok, y, 1.)
+  return jnp.where(x_ok, lax.mul(safe_x, lax.log(safe_y)), jnp.zeros_like(x))
 
 
 @_wraps(osp_special.xlog1py, update_doc=False)
 def xlog1py(x, y):
   x, y = _promote_args_inexact("xlog1py", x, y)
-  return lax._safe_mul(x, lax.log1p(y))
+  x_ok = x != 0.
+  safe_x = jnp.where(x_ok, x, 1.)
+  safe_y = jnp.where(x_ok, y, 1.)
+  return jnp.where(x_ok, lax.mul(safe_x, lax.log1p(safe_y)), jnp.zeros_like(x))
 
 
 @_wraps(osp_special.entr)
