@@ -569,7 +569,7 @@ class LaxBackedNumpyTests(jtu.JaxTestCase):
          "axis": axis, "keepdims": keepdims, "inexact": rec.inexact}
         for shape in rec.shapes for dtype in rec.dtypes
         for out_dtype in [None] + rec.dtypes
-        for axis in set(range(-len(shape), len(shape))) | set([None])
+        for axis in list(range(-len(shape), len(shape))) + [None]
         for keepdims in [False, True])
     for rec in JAX_REDUCER_RECORDS))
   def testReducer(self, onp_op, jnp_op, rng_factory, shape, dtype, out_dtype,
@@ -601,7 +601,7 @@ class LaxBackedNumpyTests(jtu.JaxTestCase):
        "axis": axis, "keepdims": keepdims, "inexact": rec.inexact}
       for rec in JAX_REDUCER_NO_DTYPE_RECORDS
       for shape in rec.shapes for dtype in rec.dtypes
-      for axis in set(range(-len(shape), len(shape))) | set([None])
+      for axis in list(range(-len(shape), len(shape))) + [None]
       for keepdims in [False, True]))
   def testReducerNoDtype(self, onp_op, jnp_op, rng_factory, shape, dtype, axis,
                          keepdims, inexact):
@@ -618,7 +618,7 @@ class LaxBackedNumpyTests(jtu.JaxTestCase):
           jtu.format_shape_dtype_string(shape, dtype), axis),
        "shape": shape, "dtype": dtype, "axis": axis}
       for shape in all_shapes for dtype in all_dtypes
-      for axis in set(range(-len(shape), len(shape))) | set([None])))
+      for axis in list(range(-len(shape), len(shape))) + [None]))
   def testCountNonzero(self, shape, dtype, axis):
     rng = jtu.rand_some_zero()
     onp_fun = lambda x: onp.count_nonzero(x, axis)
@@ -1319,6 +1319,10 @@ class LaxBackedNumpyTests(jtu.JaxTestCase):
     self._CheckAgainstNumpy(onp_op, jnp_op, args_maker, check_dtypes=True)
     self._CompileAndCheck(jnp_op, args_maker, check_dtypes=True)
 
+  def testOnesWithInvalidShape(self):
+    with self.assertRaises(TypeError):
+      jnp.ones((-1, 1))
+
   @parameterized.named_parameters(jtu.cases_from_list(
       {"testcase_name": "_inshape={}_filldtype={}_outdtype={}".format(
           jtu.format_shape_dtype_string(shape, in_dtype),
@@ -1490,7 +1494,7 @@ class LaxBackedNumpyTests(jtu.JaxTestCase):
        "rng_factory": jtu.rand_default, "shape": shape, "dtype": dtype, "axis": axis,
        "weights_shape": weights_shape, "returned": returned}
       for shape, dtype in _shape_and_dtypes(nonempty_shapes, number_dtypes)
-      for axis in set(range(-len(shape), len(shape))) | set([None])
+      for axis in list(range(-len(shape), len(shape))) + [None]
       # `weights_shape` is either `None`, same as the averaged axis, or same as
       # that of the input
       for weights_shape in ([None, shape] if axis is None or len(shape) == 1
@@ -2119,17 +2123,20 @@ class LaxBackedNumpyTests(jtu.JaxTestCase):
     jax_numpy_result = ((x + x.T) / 2).dtype
     self.assertEqual(orig_numpy_result, jax_numpy_result)
 
-  def testIssue347(self):
-    # https://github.com/google/jax/issues/347
-    def test_fail(x):
-      x = jnp.sqrt(jnp.sum(x ** 2, axis=1))
-      ones = jnp.ones_like(x)
-      x = jnp.where(x > 0.5, x, ones)
-      return jnp.sum(x)
-
-    x = jnp.array([[1, 2], [3, 4], [0, 0]], dtype=jnp.float64)
-    result = api.grad(test_fail)(x)
-    assert not onp.any(onp.isnan(result))
+  # NOTE(mattjj): I disabled this test when removing lax._safe_mul because
+  # introducing the convention 0 * inf = 0 leads to silently wrong results in
+  # some cases. See this comment for details:
+  # https://github.com/google/jax/issues/1052#issuecomment-514083352
+  # def testIssue347(self):
+  #   # https://github.com/google/jax/issues/347
+  #   def test_fail(x):
+  #     x = jnp.sqrt(jnp.sum(x ** 2, axis=1))
+  #     ones = jnp.ones_like(x)
+  #     x = jnp.where(x > 0.5, x, ones)
+  #     return jnp.sum(x)
+  #   x = jnp.array([[1, 2], [3, 4], [0, 0]], dtype=jnp.float64)
+  #   result = api.grad(test_fail)(x)
+  #   assert not onp.any(onp.isnan(result))
 
   def testIssue453(self):
     # https://github.com/google/jax/issues/453
@@ -2246,11 +2253,14 @@ class LaxBackedNumpyTests(jtu.JaxTestCase):
     self.assertAllClose(onp.zeros(3,), api.grad(f)(onp.ones(3,)),
                         check_dtypes=True)
 
-  def testIssue777(self):
-    x = jnp.linspace(-200, 0, 4, dtype=onp.float32)
-    f = api.grad(lambda x: jnp.sum(1 / (1 + jnp.exp(-x))))
-    self.assertAllClose(f(x), onp.array([0., 0., 0., 0.25], dtype=onp.float32),
-                        check_dtypes=True)
+  # NOTE(mattjj): I disabled this test when removing lax._safe_mul because this
+  # is a numerical stability issue that should be solved with a custom jvp rule
+  # of the sigmoid function being differentiated here, not by safe_mul.
+  # def testIssue777(self):
+  #   x = jnp.linspace(-200, 0, 4, dtype=onp.float32)
+  #   f = api.grad(lambda x: jnp.sum(1 / (1 + jnp.exp(-x))))
+  #   self.assertAllClose(f(x), onp.array([0., 0., 0., 0.25], dtype=onp.float32),
+  #                       check_dtypes=True)
 
   @parameterized.named_parameters(
       jtu.cases_from_list(
